moved runtime files to runtime assembly
