+++ src/mlock-unity-project/Packages/MLock/Runtime/Debugging/DebugExtensions.cs
@@ -1,13 +1,13 @@
 using System;
 using Migs.MLock.Interfaces;
 
-namespace Migs.MLock.Editor.DebugWindow
+namespace Migs.MLock.Debugging
 {
     /// <summary>
     /// Extension methods for MLock classes to easily enable debugging
     /// Follows Open/Closed principle by extending functionality without modifying original classes
     /// </summary>
-    public static class MLockDebugExtensions
+    public static class DebugExtensions
     {
         /// <summary>
         /// Registers this lock service with the debug system
@@ -18,7 +18,7 @@ namespace Migs.MLock.Editor.DebugWindow
         public static ILockService<TLockTags> WithDebug<TLockTags>(this ILockService<TLockTags> service) where TLockTags : Enum
         {
             #if UNITY_EDITOR
-            MLockDebugData.RegisterLockService(service);
+            DebugData.RegisterLockService(service);
             #endif
             
             return service;
@@ -33,22 +33,7 @@ namespace Migs.MLock.Editor.DebugWindow
         public static ILockService<TLockTags> WithoutDebug<TLockTags>(this ILockService<TLockTags> service) where TLockTags : Enum
         {
             #if UNITY_EDITOR
-            MLockDebugData.UnregisterLockService(service);
-            #endif
-            
-            return service;
-        }
-        
-        /// <summary>
-        /// Opens the MLock debug window
-        /// </summary>
-        /// <param name="service">The lock service</param>
-        /// <typeparam name="TLockTags">The enum type used for lock tags</typeparam>
-        /// <returns>The lock service for method chaining</returns>
-        public static ILockService<TLockTags> OpenDebugWindow<TLockTags>(this ILockService<TLockTags> service) where TLockTags : Enum
-        {
-            #if UNITY_EDITOR
-            MLockDebugWindow.ShowWindow();
+            DebugData.UnregisterLockService(service);
             #endif
             
             return service;
